rename currency-editor.vue to wallet-editor.vue
