+++ src/components/wallet-editor.vue
